LOG-3484: refactor package
