+++ internal/visualization/console/reconciler_test.go
@@ -1,4 +1,4 @@
-package consoleplugin
+package console
 
 import (
 	"context"
